# dist/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Caltex Music",
        "description": "",
        "contact": {
            "email": "tauovir.khan@appinventiv.com"
        },
        "version": "1.0"
    },
    "host": "caltex.appinventive.com",
    "basePath": "/api/v1",
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Customer Login",
                "description": "Login",
                "operationId": "Customer login",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Basic Auth",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "cust_email",
                        "in": "formData",
                        "description": "email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "device_id",
                        "in": "formData",
                        "description": "device id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "platform",
                        "in": "formData",
                        "description": "1=>android,2=>ios",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "device_token",
                        "in": "formData",
                        "description": "Device Token",
                        "type": "string"
                    },
                    {
                        "name": "lang",
                        "in": "formData",
                        "description": "en",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully executed"
                    },
                    "201": {
                        "description": "fail to execute"
                    },
                    "202": {
                        "description": "Require parameter missing or problem"
                    },
                    "203": {
                        "description": "Basic Authentication fail"
                    },
                    "500": {
                        "description": "Internal server Error "
                    }
                },
                "security": [
                    {
                        "basicAuth": [
                            "Username: ",
                            "Password: "
                        ]
                    }
                ]
            }
        },
...
        "/last-active-time": {
            "post": {
                "tags": [
                    "LastPresence"
                ],
                "summary": "Last Presence",
                "description": "Login",
                "operationId": "Customer login",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Basic Auth",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "active_time",
                        "in": "formData",
                        "description": "YYYY-MM-DD H:i:s",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "last_active_time",
                        "in": "formData",
                        "description": "YYYY-MM-DD H:i:s",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "device_id",
                        "in": "formData",
                        "description": "device id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "platform",
                        "in": "formData",
                        "description": "1=>android,2=>ios",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "accessToken",
                        "in": "formData",
                        "description": "Access Token",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully executed"
                    },
                    "201": {
                        "description": "fail to execute"
                    },
                    "202": {
                        "description": "Require parameter missing or problem"
                    },
                    "203": {
                        "description": "Basic Authentication fail"
                    },
                    "500": {
                        "description": "Internal server Error "
                    }
                },
                "security": [
                    {
                        "basicAuth": [
                            "Username: ",
                            "Password: "
                        ]
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "basicAuth": {
            "type": "basic"
        }
    }
}